x11-toolkits/copperspice: resurrect CopperSpice
CopperSpice is a toolkit, forked from Qt and updated to use
modern C++ and CMake in the Qt 5.something LGPL days. It was
removed from the tree for being unfetchable in 2017, now
restored. I didn't bother to look at the old ports files, so
this is entirely new work.

CS builds cleanly, except I patched in -pthread as a linker
option; I think this ought to be part of the Threads package
found by CMake, but it isn't (on FreeBSD at least). Some linkage
options need to be PUBLIC to be carried through to consuming
applications (this is a FreeBSD thing).

While here, introduce the misc/copperspice-examples which
is a demo application *kitchensink* that exercises the libraries.

CopperSpice shares notional-filenames with Qt (e.g. binaries
called "lupdate" for UI design) but the Qt ports are versioned
("lupdate-qt5"). CopperSpice gets "-cs" as a suffix.

There's a bunch of patching to make things "behave" like a
regularly packaged set of libraries and applications that
consume those libraries. In particular using $(LOCALBASE)/share/
rather than putting everything in the same target directory.